GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - master (refs/gitbutler/master)
   - src/lib/contract_abi.json
   - src/lib/contract_api.tsx
   - src/app/contracts/Records.sol
   - README.md

Your previous branch was: refs/heads/master

The sha for that commit was: 49b74531aaac4e499564fc116c270630ecff6251

For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch